move videousecase out of cameraSession so that it may persist
